[Refactor] 불필요한 코드 삭제 및 class 패키지 이동
