+++ src/main/java/com/bbangle/bbangle/board/controller/BoardController.java
@@ -1,18 +1,19 @@
 package com.bbangle.bbangle.board.controller;
 
+import com.bbangle.bbangle.board.constant.FolderBoardSortType;
+import com.bbangle.bbangle.board.constant.SortType;
 import com.bbangle.bbangle.board.dto.BoardResponse;
 import com.bbangle.bbangle.board.dto.BoardUploadRequest;
 import com.bbangle.bbangle.board.dto.FilterRequest;
 import com.bbangle.bbangle.board.facade.BoardFacade;
 import com.bbangle.bbangle.board.service.BoardService;
 import com.bbangle.bbangle.board.service.BoardUploadService;
-import com.bbangle.bbangle.board.sort.FolderBoardSortType;
-import com.bbangle.bbangle.board.sort.SortType;
 import com.bbangle.bbangle.common.dto.CommonResult;
 import com.bbangle.bbangle.common.page.CursorPageResponse;
 import com.bbangle.bbangle.common.page.CursorPagination;
 import com.bbangle.bbangle.common.service.ResponseService;
 import com.bbangle.bbangle.search.controller.mapper.SearchMapper;
+import com.bbangle.bbangle.search.facade.SearchFacade;
 import com.bbangle.bbangle.search.service.dto.SearchCommand;
 import com.bbangle.bbangle.search.service.dto.SearchInfo;
 import io.swagger.v3.oas.annotations.Operation;
@@ -21,7 +22,13 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
 @RestController
@@ -32,7 +39,8 @@ public class BoardController {
 
     private final ResponseService responseService;
     private final BoardService boardService;
-    private final RecommendBoardService recommendBoardService;
+    private final BoardFacade boardFacade;
+    private final SearchMapper searchMapper;
     private final BoardUploadService boardUploadService;
 
     @PostMapping("/board/{storeId}")
@@ -58,15 +66,6 @@ public class BoardController {
                 SearchCommand.Main command = searchMapper.toSearchMain(filterRequest, sort, null, cursorId, memberId);
                 CursorPagination<SearchInfo.Select> searchBoardPage = boardFacade.getBoardList(command);
                 return responseService.getSingleResult(searchBoardPage);
-        }
-
-        @GetMapping("/count")
-        public CommonResult getCount(
-            @ParameterObject
-            FilterRequest filterRequest
-        ) {
-                Long boardCount = boardService.getFilteredBoardCount(filterRequest);
-                return responseService.getSingleResult(boardCount);
         }
 
         @GetMapping("/folders/{folderId}")
