+++ src/main/java/com/bbangle/bbangle/board/repository/MemberSegmentRepository.java
@@ -1,6 +1,6 @@
-package com.bbangle.bbangle.board.recommend.repository;
+package com.bbangle.bbangle.board.repository;
 
-import com.bbangle.bbangle.board.recommend.domain.MemberSegment;
+import com.bbangle.bbangle.board.domain.MemberSegment;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
