+++ src/main/java/com/bbangle/bbangle/board/repository/RecommendationSimilarBoardMemoryRepository.java
@@ -1,4 +1,4 @@
-package com.bbangle.bbangle.board.repository.temp;
+package com.bbangle.bbangle.board.repository;
 
 import com.bbangle.bbangle.board.domain.RecommendBoardConfig;
 import com.bbangle.bbangle.board.domain.RedisKeyEnum;
