+++ src/main/java/com/bbangle/bbangle/search/service/dto/SearchCommand.java
@@ -1,7 +1,7 @@
 package com.bbangle.bbangle.search.service.dto;
 
 import com.bbangle.bbangle.board.dto.FilterRequest;
-import com.bbangle.bbangle.board.sort.SortType;
+import com.bbangle.bbangle.board.constant.SortType;
 
 public class SearchCommand {
