+++ src/main/java/com/bbangle/bbangle/wishlist/repository/sort/strategy/LowPriceInFolderSortRepository.java
@@ -4,7 +4,7 @@ import static com.bbangle.bbangle.board.domain.QBoard.board;
 import static com.bbangle.bbangle.board.repository.BoardRepositoryImpl.BOARD_PAGE_SIZE;
 import static com.bbangle.bbangle.wishlist.domain.QWishListBoard.wishListBoard;
 
-import com.bbangle.bbangle.board.sort.FolderBoardSortType;
+import com.bbangle.bbangle.board.constant.FolderBoardSortType;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.NumberPath;
